fix(templates): tarball to zip
semantic-release/exec uses lodash when parsing the exec commands.
Since lodash by default supports the ES6 delimiter and
semantic-release uses it that default way, we are forced to not use
terms that look like `${...}` in our commands. The fix is to use sed
instead to get rid of the tarball file extension when replacing it
with the zip extension.